backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
python-dotenv
bcrypt==4.0.1
passlib[bcrypt]
python-jose[cryptography]
